[software] Split Track and TracksBuilder in two different files
