+++ src/aliceVision/sfm/pipeline/panorama/ReconstructionEngine_panorama.cpp
@@ -14,7 +14,7 @@
 #include <aliceVision/system/Timer.hpp>
 #include <aliceVision/stl/stl.hpp>
 #include <aliceVision/multiview/essential.hpp>
-#include <aliceVision/track/Track.hpp>
+#include <aliceVision/track/TracksBuilder.hpp>
 #include <aliceVision/config.hpp>
