+++ src/aliceVision/track/TracksBuilder.cpp
@@ -5,7 +5,7 @@
 // v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "Track.hpp"
+#include "TracksBuilder.hpp"
 
 namespace aliceVision {
 namespace track {
